Crate tower_cache

Cache layer for tower::Services

CacheLayer is a tower Layer that provides caches for Services by using another service to handle the cache. This allows the usage of asynchronous and external caches.

Usage

Here is a sample implementation using an LRU cache as cache provider:

use std::convert::Infallible;
use tower::{Service, ServiceBuilder, service_fn};
use tower_cache::{
    CacheLayer,
    lru::LruProvider,
};
async fn handler(req: String) -> Result<String, Infallible> {
    Ok(req.to_uppercase())
}

// Initialize the cache provider service
let lru_provider = LruProvider::new::<String, String>(20);

// Wrap the service with CacheLayer.
let mut my_service = ServiceBuilder::new()
    .layer(CacheLayer::new(lru_provider))
    .service(service_fn(handler));

// Call the service
let res = my_service.call("Hello".to_string()).await.unwrap();
assert_eq!(res, "HELLO".to_string());

// Calling the service again will use the cache instead
let res = my_service.call("Hello".to_string()).await.unwrap();
assert_eq!(res, "HELLO".to_string());

With request transformer

Certain cache providers might require specific trait bounds to be met in order to cache that data. Or you might want to cache based on specific fields in a struct, such as a request ID, URI path, etc.

For this reason, you can specify a transformer function, that will transform incoming requests before sending them to the cache provider.

use std::convert::Infallible;
use tower::{Service, ServiceBuilder, service_fn};
use tower_cache::{
    CacheLayer,
    lru::LruProvider,
};
async fn handler(req: String) -> Result<String, Infallible> {
    Ok(req.to_uppercase())
}
 
fn transform_req(req: String) -> usize {
    req.len()
}

// Initialize the cache provider service
let lru_provider = LruProvider::new::<usize, String>(20);
 
let cache_layer = CacheLayer::new(lru_provider)
    .with_transformer(transform_req);

// Wrap the service with CacheLayer.
let mut my_service = ServiceBuilder::new()
    .layer(cache_layer)
    .service(service_fn(handler));

// Call the service
let res = my_service.call("Hello".to_string()).await.unwrap();
assert_eq!(res, "HELLO".to_string());
 
// Since this uses a transformer that takes the length of the String,
// we will get the same result for a string of similar length.
let res = my_service.call("Salut".to_string()).await.unwrap();
assert_eq!(res, "HELLO".to_string());

Creating cache providers

A cache provider is a tower::Service that takes a ProviderRequest as request and returns a ProviderResponse.

Modules

lru

LRU cache provider

Structs

CacheLayer

Layer that adds cache to a tower::Service

CacheService

Service generated by CacheLayer.

Enums

Error

Error returned by the CacheLayer

ProviderRequest

Requests sent to the cache provider

ProviderResponse

Responses sent by the cache provider

Traits

Transform

Request transformation trait