Trait tower_cache::Transform

pub trait Transform<R> {
    type Output;
    fn transform(&self, req: R) -> Self::Output;
}

Request transformation trait

In many cases, it’s not useful to cache based on the entire request payload, or the request might not be directly usable with the desired cache provider. This trait provide an abstraction over functions that transform a request into a value that can be used by a caching provider.

Usage

In most case, you don’t need to implement this trait directly. It is automatically implemented for functions that take one argument and return another.

use tower_cache::Transform;

fn my_func(req: usize) -> usize {
    req * 2
}

// Calling my_func using the Transform trait
assert_eq!((my_func).transform(2), 4);

This is also implemented for () as a no-op transformer:

use tower_cache::Transform;

assert_eq!(().transform(2), 2);

Associated Types

type Output

Output of the transformer

Required methods

fn transform(&self, req: R) -> Self::Output

Transform a key into a reference value for a cache provider.